Клиент и docker-compose.yml
