+++ docker-compose.yml
@@ -1,14 +1,35 @@
 services:
   stats-server:
+    container_name: stats-server
+    build: stats-service/stats-server
     ports:
       - "9090:9090"
+    depends_on:
+      - stats-db
+    environment:
+      - SPRING_DATASOURCE_URL=jdbc:postgresql://stats-db-postgres:5432/ewm-stats-db
+      - SPRING_DATASOURCE_USERNAME=dbuser
+      - SPRING_DATASOURCE_PASSWORD=12345
 
   stats-db:
+    container_name: stats-db-postgres
     image: postgres:16.1
-
-  ewm-service:
     ports:
-      - "8080:8080"
+      - "6432:5432"
+    environment:
+      - POSTGRES_DB=ewm-stats-db
+      - POSTGRES_USER=dbuser
+      - POSTGRES_PASSWORD=12345
+    healthcheck:
+      test: pg_isready -q -d $$POSTGRES_DB -U $$POSTGRES_USER
+      timeout: 5s
+      interval: 5s
+      retries: 10
 
-  ewm-db:
-    image: postgres:16.1
+
+#  ewm-service:
+#    ports:
+#      - "8080:8080"
+
+#  ewm-db:
+#    image: postgres:16.1
