+++ stats-service/pom.xml
@@ -13,8 +13,8 @@
     <packaging>pom</packaging>
 
     <modules>
-        <module>stats-client</module>
         <module>stats-dto</module>
+        <module>stats-client</module>
         <module>stats-server</module>
     </modules>
 
@@ -31,5 +31,4 @@
             <optional>true</optional>
         </dependency>
     </dependencies>
-
 </project>
