+++ stats-service/stats-client/src/main/java/TestClient.java
@@ -8,12 +8,9 @@ import java.util.List;
 public class TestClient {
 
     public static void main(String[] args) {
-
         StatsClient sc = new StatsClientImpl("http://localhost:9090");
         HitDto hit = sc.addHit("3112", "/rrr/tttt", "10.10.10.1", LocalDateTime.now());
-
         List<String> uris = List.of("/rrr/tttt", "/343434/ffff");
-
         System.out.println("hit = " + hit);
         System.out.println(
                 "stats = " + sc.getStats(LocalDateTime.now().minusDays(1),
