+++ stats-service/stats-client/src/main/java/ru/practicum/emw/stats/StatsClientImpl.java
@@ -3,7 +3,6 @@ package ru.practicum.emw.stats;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import org.springframework.web.client.RestClient;
 import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponentsBuilder;
 import ru.practicum.ewm.stats.dto.HitDto;
@@ -11,18 +10,19 @@ import ru.practicum.ewm.stats.dto.StatsDto;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 @Component
 public class StatsClientImpl implements StatsClient {
 
-    //private final RestClient restClient;
     private final RestTemplate restTemplate;
-
     private static final String HIT_ENDPOINT = "/hit";
     private static final String STATS_ENDPOINT = "/stats";
     private static final DateTimeFormatter dtf = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
-    private String baseUrl;
+    private final String baseUrl;
+
 
     @Autowired
     public StatsClientImpl(@Value("${stats-server.url}") String statsServerUrl) {
@@ -48,11 +48,11 @@ public class StatsClientImpl implements StatsClient {
 
         String resultUrl = builder.build().toUriString();
 
-        StatsDto[] statsArray =  restTemplate.getForObject(baseUrl + resultUrl, StatsDto[].class);
+        StatsDto[] statsArray = restTemplate.getForObject(baseUrl + resultUrl, StatsDto[].class);
         if (statsArray != null) {
             return Arrays.asList(statsArray);
         } else {
-            return new ArrayList<StatsDto>();
+            return new ArrayList<>();
         }
     }
 }
