+++ stats-service/stats-dto/pom.xml
@@ -8,7 +8,6 @@
         <artifactId>stats-service</artifactId>
         <version>0.0.1-SNAPSHOT</version>
     </parent>
-
     <artifactId>stats-dto</artifactId>
 
     <properties>
