+++ stats-service/stats-dto/src/main/java/ru/practicum/ewm/stats/mapper/HitMapper.java
@@ -2,7 +2,6 @@ package ru.practicum.ewm.stats.mapper;
 
 
 import org.mapstruct.Mapper;
-import org.mapstruct.Mapping;
 import ru.practicum.ewm.stats.dto.HitDto;
 import ru.practicum.ewm.stats.model.Hit;
