+++ stats-service/stats-server/dockerfile
@@ -1,0 +1,3 @@
+FROM amazoncorretto:21-alpine
+COPY target/*.jar app.jar
+ENTRYPOINT ["java","-jar","/app.jar"]
