+++ stats-service/stats-server/src/main/java/ru/practicum/ewm/stats/repository/StatsRepository.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewm.stats;
+package ru.practicum.ewm.stats.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
@@ -16,19 +16,19 @@ public interface StatsRepository extends JpaRepository<Hit, Long> {
     //         "or lower(it.description) like lower(concat('%', :text,'%')) order by it.id")
     // List<Item> findAvailableByNameOrDescription(@Param("text") String text);
 
-     @Query("select new ru.practicum.ewm.stats.model.Stats(h.app, h.uri, COUNT(h.ip))" +
+    @Query("select new ru.practicum.ewm.stats.model.Stats(h.app, h.uri, COUNT(h.ip))" +
             " FROM Hit AS h where h.timestamp between :start and :end" +
             " GROUP BY h.app, h.uri ORDER BY COUNT(h.ip) DESC"
-            )
+    )
     List<Stats> getAllHits(@Param("start") LocalDateTime start,
-                                 @Param("end") LocalDateTime end);
+                           @Param("end") LocalDateTime end);
 
     @Query("select new ru.practicum.ewm.stats.model.Stats(h.app, h.uri, COUNT(DISTINCT h.ip))" +
             " FROM Hit AS h where h.timestamp between :start and :end" +
             " GROUP BY h.app, h.uri ORDER BY COUNT(DISTINCT h.ip) DESC"
     )
     List<Stats> getAllUniqueHits(@Param("start") LocalDateTime start,
-                           @Param("end") LocalDateTime end);
+                                 @Param("end") LocalDateTime end);
 
     @Query("select new ru.practicum.ewm.stats.model.Stats(h.app, h.uri, COUNT(DISTINCT h.ip))" +
             " FROM Hit AS h where (h.timestamp between :start and :end) and h.uri IN :uriList" +
@@ -43,7 +43,7 @@ public interface StatsRepository extends JpaRepository<Hit, Long> {
             " GROUP BY h.app, h.uri ORDER BY COUNT(h.ip) DESC"
     )
     List<Stats> getAllHitsByUris(@Param("start") LocalDateTime start,
-                                       @Param("end") LocalDateTime end,
-                                       @Param("uriList") List<String> uris);
+                                 @Param("end") LocalDateTime end,
+                                 @Param("uriList") List<String> uris);
 
 }
