+++ stats-service/stats-server/src/main/java/ru/practicum/ewm/stats/service/StatsService.java
@@ -1,4 +1,4 @@
-package ru.practicum.ewm.stats;
+package ru.practicum.ewm.stats.service;
 
 
 import ru.practicum.ewm.stats.dto.HitDto;
